+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `file_ttl` support to Secure Share.
+
 ## 4.2.0 - 2024-12-18
 
 ### Added