+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 4.3.0 - 2025-01-13
+
 ### Added
 
 - `file_ttl` support to Secure Share.